+++ src/views/tool/schema/index.vue
@@ -37,14 +37,23 @@
           :height="height - 120"
           :style="{ overflowY: 'auto' }"
         >
-          <JsonSchema
-            ref="jsonSchema"
-            :app-values="appValues"
-            class="px-2 pt-3"
-            :cluster-name="''"
-            :params="params"
-            @changeBasicFormParam="changeBasicFormParam"
-          />
+          <v-form ref="form" v-model="valid" class="pa-0 ma-3" lazy-validation @submit.prevent>
+            <v-flex class="pa-0 ma-0">
+              <BaseParam
+                v-for="(param, index) in params"
+                :id="`p${index}`"
+                :key="`p${index}`"
+                :all-params="params"
+                :app-values="appValues"
+                class="mt-0"
+                :cluster-name="''"
+                :param="param"
+                v-bind="$attrs"
+                v-on="$listeners"
+                @changeBasicFormParam="changeBasicFormParam"
+              />
+            </v-flex>
+          </v-form>
         </v-card>
       </v-col>
     </v-row>
@@ -55,14 +64,11 @@
   import { mapState } from 'vuex';
 
   import { deepCopy } from '@/utils/helpers';
-  import { getValueSchema, setValue } from '@/utils/yaml';
-  import JsonSchema from '@/views/appstore/components/DeployWizard/JsonSchema';
+  import { retrieveFromSchema } from '@/utils/schema';
+  import { setValue } from '@/utils/yaml';
 
   export default {
     name: 'SchemaForm',
-    components: {
-      JsonSchema,
-    },
     data() {
       return {
         appValues: {},
@@ -78,7 +84,7 @@
     },
     methods: {
       validate() {
-        return this.$refs.jsonSchema.validate();
+        return this.$refs.form.validate();
       },
       getData() {
         this.obj.values = deepCopy(this.appValues);
@@ -87,7 +93,7 @@
       async renderSchema() {
         this.appValues = {};
         this.schemaJson = JSON.parse(this.jsonSchema);
-        this.params = this.retrieveBasicFormParams(this.appValues, this.schemaJson);
+        this.params = retrieveFromSchema(this.appValues, this.schemaJson);
       },
       changeBasicFormParam(param, value) {
         // Change raw values 修改原始值, 返回的是字符串
@@ -96,46 +102,7 @@
       },
       reRender() {
         this.params = [];
-        this.params = this.retrieveBasicFormParams(this.appValues, this.schemaJson);
-      },
-      retrieveBasicFormParams(defaultValues, schema, parentPath = '') {
-        let params = [];
-        if (schema && schema.properties) {
-          const properties = schema.properties;
-          Object.keys(properties).forEach((propertyKey) => {
-            const itemPath = `${parentPath || ''}${propertyKey}`;
-            const { type, form } = properties[propertyKey];
-            if (form) {
-              // Use the default value either from the JSON schema or the default values
-              // 使用schema中的默认值
-              // const value = properties[propertyKey].default
-              // 使用values.yaml的默认值
-              const value = getValueSchema(defaultValues, itemPath, properties[propertyKey].default);
-              const param = {
-                ...properties[propertyKey],
-                path: itemPath,
-                name: propertyKey,
-                type,
-                value,
-                enum: properties[propertyKey].enum?.map((item) => item?.toString() ?? ''),
-                children:
-                  properties[propertyKey].type === 'object'
-                    ? this.retrieveBasicFormParams(defaultValues, properties[propertyKey], `${itemPath}/`)
-                    : undefined,
-              };
-              params = params.concat(param);
-            } else {
-              // form为假不渲染
-              // If the property is an object, iterate recursively 递归遍历
-              if (schema.properties[propertyKey].type !== 'object') {
-                params = params.concat(
-                  this.retrieveBasicFormParams(defaultValues, properties[propertyKey], `${itemPath}/`),
-                );
-              }
-            }
-          });
-        }
-        return params;
+        this.params = retrieveFromSchema(this.appValues, this.schemaJson);
       },
       reset() {
         this.appValues = {};